Python/E11_Metadata_Imagenes_Web/requirements.txt
lxml==4.6.3
requests==2.25.1
beautifulsoup4==4.10.0
Pillow==8.3.2
